Create main.py
Simple Telegram Bot (which can be hosted on Pythonanywhere) that answers questions in any language on this document by the Geneva Hospice General: https://www.hospicegeneral.ch/sites/default/files/Directives_version_validee_DSE.pdf

It tries to answer any question on this topic in Ukrainian. However, this is NOT an official bot by the Hospice General. Any info you take from this bot should be confirmed directly by getting in contact with the Hospice General.